Rename LocationTag.interface.ts to LocationTag.ts
